\name{PeMSD3}
\alias{PeMSD3}
\docType{data}
\title{
PeMSD3
}
\description{
The PeMSD3 data
}
\usage{data("PeMSD3")}
\format{
  A data frame with 26208 observations on the following 12 variables.
  \describe{
    \item{\code{X315836}}{a numeric vector}
    \item{\code{X315837}}{a numeric vector}
    \item{\code{X315838}}{a numeric vector}
    \item{\code{X315841}}{a numeric vector}
    \item{\code{X315842}}{a numeric vector}
    \item{\code{X315839}}{a numeric vector}
    \item{\code{X315843}}{a numeric vector}
    \item{\code{X315846}}{a numeric vector}
    \item{\code{X315847}}{a numeric vector}
    \item{\code{X317895}}{a numeric vector}
    \item{\code{X315849}}{a numeric vector}
    \item{\code{X315850}}{a numeric vector}
  }
}
\details{
It is the traffic data of Sacramento in California, the United States.
}
\source{
Song, C., Lin, Y., Guo, S., Wan, H. Spatial-temporal synchronous graph convolutional networks: a new framework for spatial-temporal network data forecasting[C]. Proceedings of the AAAI Conference on Artificial Intelligence, 34(1), 914-921.
}
\references{
Song, C., Lin, Y., Guo, S., Wan, H. Spatial-temporal synchronous graph convolutional networks: a new framework for spatial-temporal network data forecasting[C]. Proceedings of the AAAI Conference on Artificial Intelligence, 34(1), 914-921.
}
\examples{
data(PeMSD3)
## maybe str(PeMSD3) ; plot(PeMSD3) ...
}
\keyword{datasets}
