\name{PeMSD7}
\alias{PeMSD7}
\docType{data}
\title{
PeMSD7
}
\description{
The PeMSD7 data
}
\usage{data("PeMSD7")}
\format{
  A data frame with 17568 observations on the following 20 variables.
  \describe{
    \item{\code{X773656}}{a numeric vector}
    \item{\code{X760074}}{a numeric vector}
    \item{\code{X760080}}{a numeric vector}
    \item{\code{X716414}}{a numeric vector}
    \item{\code{X760101}}{a numeric vector}
    \item{\code{X760112}}{a numeric vector}
    \item{\code{X716419}}{a numeric vector}
    \item{\code{X716421}}{a numeric vector}
    \item{\code{X716424}}{a numeric vector}
    \item{\code{X765476}}{a numeric vector}
    \item{\code{X760167}}{a numeric vector}
    \item{\code{X716427}}{a numeric vector}
    \item{\code{X716431}}{a numeric vector}
    \item{\code{X716433}}{a numeric vector}
    \item{\code{X760187}}{a numeric vector}
    \item{\code{X760196}}{a numeric vector}
    \item{\code{X716440}}{a numeric vector}
    \item{\code{X760226}}{a numeric vector}
    \item{\code{X760236}}{a numeric vector}
    \item{\code{X716449}}{a numeric vector}
  }
}
\details{
It is the traffic data of Los Angeles in California, the United States.
}
\source{
Xu, D., Wei, C., Peng, P., Xuan, Q., Guo, H. Ge-gan: a novel deep learning framework for road traffic state estimation[J]. Transportation Research Part C Emerging Technologies, 2020, 117, 102635.
}
\references{
Xu, D., Wei, C., Peng, P., Xuan, Q., Guo, H. Ge-gan: a novel deep learning framework for road traffic state estimation[J]. Transportation Research Part C Emerging Technologies, 2020, 117, 102635.
}
\examples{
data(PeMSD7)
## maybe str(PeMSD7) ; plot(PeMSD7) ...
}
\keyword{datasets}
