\name{PeMSD7_weekend}
\alias{PeMSD7_weekend}
\docType{data}
\title{
PeMSD7 on weekend
}
\description{
The weekend data in PeMSD7
}
\usage{data("PeMSD7_weekend")}
\format{
  A data frame with 5184 observations on the following 20 variables.
  \describe{
    \item{\code{X773656}}{a numeric vector}
    \item{\code{X760074}}{a numeric vector}
    \item{\code{X760080}}{a numeric vector}
    \item{\code{X716414}}{a numeric vector}
    \item{\code{X760101}}{a numeric vector}
    \item{\code{X760112}}{a numeric vector}
    \item{\code{X716419}}{a numeric vector}
    \item{\code{X716421}}{a numeric vector}
    \item{\code{X716424}}{a numeric vector}
    \item{\code{X765476}}{a numeric vector}
    \item{\code{X760167}}{a numeric vector}
    \item{\code{X716427}}{a numeric vector}
    \item{\code{X716431}}{a numeric vector}
    \item{\code{X716433}}{a numeric vector}
    \item{\code{X760187}}{a numeric vector}
    \item{\code{X760196}}{a numeric vector}
    \item{\code{X716440}}{a numeric vector}
    \item{\code{X760226}}{a numeric vector}
    \item{\code{X760236}}{a numeric vector}
    \item{\code{X716449}}{a numeric vector}
  }
}
\details{
The weekend data set only records traffic flow data on weekends.
}
\source{
Xu, D., Wei, C., Peng, P., Xuan, Q., Guo, H. Ge-gan: a novel deep learning framework for road traffic state estimation[J]. Transportation Research Part C Emerging Technologies, 2020, 117, 102635.
}
\references{
Xu, D., Wei, C., Peng, P., Xuan, Q., Guo, H. Ge-gan: a novel deep learning framework for road traffic state estimation[J]. Transportation Research Part C Emerging Technologies, 2020, 117, 102635.
}
\examples{
data(PeMSD7_weekend)
## maybe str(PeMSD7_weekend) ; plot(PeMSD7_weekend) ...
}
\keyword{datasets}
